+++ typescript/registry/src/funcs/serversList.ts
@@ -11,7 +11,6 @@ import { safeParse } from "../lib/schemas.js";
 import { RequestOptions } from "../lib/sdks.js";
 import { extractSecurity, resolveGlobalSecurity } from "../lib/security.js";
 import { pathToFunc } from "../lib/url.js";
-import { APIError } from "../models/errors/apierror.js";
 import {
   ConnectionError,
   InvalidRequestError,
@@ -20,7 +19,9 @@ import {
   UnexpectedClientError,
 } from "../models/errors/httpclienterrors.js";
 import * as errors from "../models/errors/index.js";
+import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
+import { SmitheryRegistryError } from "../models/errors/smitheryregistryerror.js";
 import * as operations from "../models/operations/index.js";
 import { APICall, APIPromise } from "../types/async.js";
 import { Result } from "../types/fp.js";
@@ -47,13 +48,14 @@ export function serversList(
       operations.ListServersResponse,
       | errors.UnauthorizedError
       | errors.ServerError
-      | APIError
-      | SDKValidationError
-      | UnexpectedClientError
-      | InvalidRequestError
+      | SmitheryRegistryError
+      | ResponseValidationError
+      | ConnectionError
       | RequestAbortedError
       | RequestTimeoutError
-      | ConnectionError
+      | InvalidRequestError
+      | UnexpectedClientError
+      | SDKValidationError
     >,
     { page: number }
   >
@@ -76,13 +78,14 @@ async function $do(
         operations.ListServersResponse,
         | errors.UnauthorizedError
         | errors.ServerError
-        | APIError
-        | SDKValidationError
-        | UnexpectedClientError
-        | InvalidRequestError
+        | SmitheryRegistryError
+        | ResponseValidationError
+        | ConnectionError
         | RequestAbortedError
         | RequestTimeoutError
-        | ConnectionError
+        | InvalidRequestError
+        | UnexpectedClientError
+        | SDKValidationError
       >,
       { page: number }
     >,
@@ -117,6 +120,7 @@ async function $do(
   const requestSecurity = resolveGlobalSecurity(securityInput);
 
   const context = {
+    options: client._options,
     baseURL: options?.serverURL ?? client._baseURL ?? "",
     operationID: "listServers",
     oAuth2Scopes: [],
@@ -148,6 +152,7 @@ async function $do(
     headers: headers,
     query: query,
     body: body,
+    userAgent: client._options.userAgent,
     timeoutMs: options?.timeoutMs || client._options.timeoutMs || -1,
   }, options);
   if (!requestRes.ok) {
@@ -174,13 +179,14 @@ async function $do(
     operations.ListServersResponse,
     | errors.UnauthorizedError
     | errors.ServerError
-    | APIError
-    | SDKValidationError
-    | UnexpectedClientError
-    | InvalidRequestError
+    | SmitheryRegistryError
+    | ResponseValidationError
+    | ConnectionError
     | RequestAbortedError
     | RequestTimeoutError
-    | ConnectionError
+    | InvalidRequestError
+    | UnexpectedClientError
+    | SDKValidationError
   >(
     M.json(200, operations.ListServersResponse$inboundSchema, {
       key: "Result",
@@ -189,7 +195,7 @@ async function $do(
     M.jsonErr(500, errors.ServerError$inboundSchema),
     M.fail("4XX"),
     M.fail("5XX"),
-  )(response, { extraFields: responseFields });
+  )(response, req, { extraFields: responseFields });
   if (!result.ok) {
     return [haltIterator(result), {
       status: "complete",
@@ -206,13 +212,14 @@ async function $do(
         operations.ListServersResponse,
         | errors.UnauthorizedError
         | errors.ServerError
-        | APIError
-        | SDKValidationError
-        | UnexpectedClientError
-        | InvalidRequestError
+        | SmitheryRegistryError
+        | ResponseValidationError
+        | ConnectionError
         | RequestAbortedError
         | RequestTimeoutError
-        | ConnectionError
+        | InvalidRequestError
+        | UnexpectedClientError
+        | SDKValidationError
       >
     >;
     "~next"?: { page: number };