+++ typescript/registry/src/lib/config.ts
@@ -30,6 +30,10 @@ export type SDKOptions = {
    */
   serverURL?: string | undefined;
   /**
+   * Allows overriding the default user agent used by the SDK
+   */
+  userAgent?: string | undefined;
+  /**
    * Allows overriding the default retry config used by the SDK
    */
   retryConfig?: RetryConfig;
@@ -57,7 +61,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.3.7",
-  genVersion: "2.604.4",
-  userAgent: "speakeasy-sdk/typescript 0.3.7 2.604.4 1.0.0 @smithery/registry",
+  sdkVersion: "0.4.0",
+  genVersion: "2.656.3",
+  userAgent: "speakeasy-sdk/typescript 0.4.0 2.656.3 1.0.0 @smithery/registry",
 } as const;