+++ typescript/registry/src/models/errors/index.ts
@@ -5,6 +5,7 @@
 export * from "./apierror.js";
 export * from "./getserver.js";
 export * from "./httpclienterrors.js";
+export * from "./responsevalidationerror.js";
 export * from "./sdkvalidationerror.js";
 export * from "./servererror.js";
 export * from "./unauthorizederror.js";